NetworkAnalyzer/requirements.txt
psutil
pandas
matplotlib
numpy
schedule
speedtest-cli
pyinstaller
PyQt5
nlohmann-json
